# apps/api/openapi.yaml
openapi: "3.0.3"
info:
  title: One Step Ahead API
  description: >
    Multiplayer step-tracking game API. Behavior is normative in
    docs/implementation-plan.md §3; this document is normative for shapes.
    All authenticated routes use the `sc_session` JWT cookie set by the
    Google OAuth callback (or `/api/auth/dev-login` outside production).
  version: "1.0.0"
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - sessionCookie: []
paths:
...
  /api/groups:
    post:
      summary: Create a group
      description: Creator becomes admin and member; the first week is created immediately.
      tags: [groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 40 }
      responses:
        "201":
          description: The new group (includes the invite code).
          content:
            application/json:
              schema:
                type: object
                properties:
                  group: { $ref: "#/components/schemas/Group" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /api/groups/join:
    post:
      summary: Join a group by invite code
      tags: [groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code]
              properties:
                invite_code:
                  type: string
                  pattern: "^[ABCDEFGHJKMNPQRSTUVWXYZ23456789]{6}$"
      responses:
        "200":
          description: Joined.
          content:
            application/json:
              schema:
                type: object
                properties:
                  group: { $ref: "#/components/schemas/Group" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: GROUP_FULL — the group already has 8 players.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /api/groups/me:
    get:
      summary: My group and members
      tags: [groups]
      responses:
        "200":
          description: Group with member list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  group: { $ref: "#/components/schemas/Group" }
                  members:
                    type: array
                    items: { $ref: "#/components/schemas/Member" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/groups/me/leave:
    post:
      summary: Leave my group
      description: >
        Dissolves the leaver's nemesis matchup (opponent re-pairs with the bye
        or becomes the bye). If the admin leaves, the oldest remaining member
        is promoted.
      tags: [groups]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/groups/me/members/{userId}:
    delete:
      summary: Remove a member (admin only)
      description: Same nemesis re-pair semantics as leaving.
      tags: [groups]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: Admin cannot remove themselves — use leave.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  # ---- Weeks ----
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sc_session
      description: JWT session cookie.
  responses:
    Ok:
      description: Success.
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean }
    Unauthenticated:
      description: Missing or invalid session.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: FORBIDDEN.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: NOT_FOUND.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    Member:
      type: object
      description: Public member projection embedded in many payloads.
      properties:
        user_id: { type: string, format: uuid }
        display_name: { type: string }
        avatar_skin: { type: integer }
        avatar_hair: { type: integer }
        avatar_colorway: { type: integer }
    Group:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        invite_code: { type: string }
        admin_id: { type: string, format: uuid }
        timezone: { type: string }
        created_at: { type: string, format: date-time }
...